requirements.txt
aif360
altair
annotated-types
anyio
attrs
audit-AI
blinker
cachetools
certifi
charset-normalizer
click
cloudpickle
contourpy
cycler
defusedxml
exceptiongroup
fastapi
fonttools
fpdf2
gitdb
GitPython
google-genai==1.73.1
google-cloud-vision==3.13.0
google-cloud-aiplatform==1.148.1
# google-generativeai is superseded by google-genai in aiplatform>=1.70
# ── Pinned transitive deps (prevents pip resolution-too-deep) ─────────────────
google-api-core==2.30.3
google-auth==2.49.0
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.80.0
grpc-google-iam-v1==0.14.4
google-cloud-storage==3.10.1
google-cloud-core==2.5.1
google-cloud-bigquery==3.41.0
google-cloud-resource-manager==1.17.0
google-crc32c==1.8.0
google-resumable-media==2.8.2
# google-ai-generativelanguage pinned by google-genai automatically
proto-plus==1.27.2
protobuf==6.31.1
h11
idna
importlib_resources
Jinja2
joblib
jsonschema-specifications
jsonschema
kiwisolver
llvmlite
MarkupSafe
matplotlib
narwhals
numba
numpy
packaging
pandas
patsy
pillow
plotly
pyarrow
pydantic
pydantic_core
pydeck
pyparsing
python-dateutil
python-dotenv
python-multipart
pytz
referencing
requests
rpds-py
scikit-learn
scipy
shap
six
slicer
smmap
starlette
statsmodels
streamlit
tenacity
threadpoolctl
toml
tornado
tqdm
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
watchdog
zipp
